extract dynamic data units from reference view module for reusability
